% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cerf_gp}
\alias{autoplot.cerf_gp}
\title{A helper function for cerf_gp object}
\usage{
\method{autoplot}{cerf_gp}(object, ...)
}
\arguments{
\item{object}{A cerf_gp object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cerf_gp object using ggplot2 package.
}
\keyword{internal}
