% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.cerf_nngp}
\alias{autoplot.cerf_nngp}
\title{A helper function for cerf_nngp object}
\usage{
\method{autoplot}{cerf_nngp}(object, ...)
}
\arguments{
\item{object}{A cerf_nngp object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot object.
}
\description{
A helper function to plot cerf_nngp object using ggplot2 package.
}
\keyword{internal}
