% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_inverse.R
\name{compute_inverse}
\alias{compute_inverse}
\title{Compute matrix inverse for a covariate matrix}
\usage{
compute_inverse(mtrx)
}
\arguments{
\item{mtrx}{An \code{n * n} covariate matrix}
}
\value{
A matrix that represent the inverse of the input matrix.
}
\description{
Computes inverse of a covariate matrix using Choleski decomposition.
}
\keyword{internal}
