% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise_gp.R
\name{estimate_noise_gp}
\alias{estimate_noise_gp}
\title{Estimate the standard deviation of the nugget term in standard Gaussian
process}
\usage{
estimate_noise_gp(data, sigma_obs, inv_sigma_obs)
}
\arguments{
\item{data}{A vector of outcome data.}

\item{sigma_obs}{Covariance matrix between observed covariates.}

\item{inv_sigma_obs}{Inverse of the covariance matrix between observed
covariates.}
}
\value{
A scalar of estimated standard deviation of the nugget term in standard GP.
}
\description{
Estimates the standard deviations of the nugget term in standard GP by
calculating the standard deviations of the residuals.
}
\keyword{internal}
