% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcerf_package.R
\docType{package}
\name{GPCERF-package}
\alias{GPCERF-package}
\alias{GPCERF}
\title{The 'GPCERF' package.}
\description{
Provides a non-parametric Bayesian framework based on Gaussian process priors
for estimating causal effects of a continuous exposure and detecting change
points in the causal exposure response curves using observational data.
}
\references{
Ren, B., Wu, X., Braun, D., Pillai, N. and Dominici, F., 2021. Bayesian
modeling for exposure response curve via gaussian processes: Causal effects
of exposure to air pollution on health outcomes. arXiv preprint
arXiv:2105.03454.
}
\author{
Naeem Khoshnevis

Boyu Ren

Danielle Braun
}
