% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cerf_gp}
\alias{plot.cerf_gp}
\title{Extend generic plot functions for cerf_gp class}
\usage{
\method{plot}{cerf_gp}(x, ...)
}
\arguments{
\item{x}{A cerf_gp object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side
effects.
}
\description{
A wrapper function to extend generic plot functions for cerf_gp class.
}
