% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cerf_nngp}
\alias{plot.cerf_nngp}
\title{Extend generic plot functions for cerf_nngp class}
\usage{
\method{plot}{cerf_nngp}(x, ...)
}
\arguments{
\item{x}{A cerf_nngp object.}

\item{...}{Additional arguments passed to customize the plot.}
}
\value{
Returns a ggplot2 object, invisibly. This function is called for side
effects.
}
\description{
A wrapper function to extend generic plot functions for cerf_nngp class.
}
