% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cerf_gp}
\alias{print.cerf_gp}
\title{Extend print function for cerf_gp object}
\usage{
\method{print}{cerf_gp}(x, ...)
}
\arguments{
\item{x}{A cerf_gp object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for cerf_gp object
}
