% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cerf_nngp}
\alias{print.cerf_nngp}
\title{Extend print function for cerf_nngp object}
\usage{
\method{print}{cerf_nngp}(x, ...)
}
\arguments{
\item{x}{A cerf_nngp object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
No return value. This function is called for side effects.
}
\description{
Extend print function for cerf_nngp object
}
