% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cerf_gp}
\alias{summary.cerf_gp}
\title{print summary of cerf_gp object}
\usage{
\method{summary}{cerf_gp}(object, ...)
}
\arguments{
\item{object}{A cerf_gp object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data
}
\description{
print summary of cerf_gp object
}
