% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.cerf_nngp}
\alias{summary.cerf_nngp}
\title{print summary of cerf_nngp object}
\usage{
\method{summary}{cerf_nngp}(object, ...)
}
\arguments{
\item{object}{A cerf_nngp object.}

\item{...}{Additional arguments passed to customize the results.}
}
\value{
Returns summary of data.
}
\description{
print summary of cerf_nngp object
}
